+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -846,6 +846,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Changed the default exposure mode to "Automatic (Histogram)", along with "Limit Min" to -4 and "Limit Max" to 16.
 - Replaced the default scene system with the builtin Scene Template feature.
 - Changed extensions of shader CAS include files.
+- Making the planar probe atlas's format match the color buffer's format.
+- Removing the planarReflectionCacheCompressed setting from asset.
 
 ## [7.1.1] - 2019-09-05